Use nanosleep instead of usleep for delays
After downloading the dfu-util 0.7 binary distribution for Windows
(MinGW), I noticed that the milli_sleep(2000) delay after a detach was
not delaying at all. Grabbing the source, and doing some debugging, I
found that MinGW now provides a usleep() implementation (HAVE_USLEEP
gets defined by configure script), but it will just return immediately
if the specified delay is 1000000 or greater. According to the usleep
man page, operation is only defined in the range [0, 1000000], with
behaviour outside of this range being undefined, so TECHNICALLY it's not
a bug.

Anyways, my "quick hack" was to just swap the order of the HAVE_USLEEP
and HAVE_WINDOWS_H checks in src/portable.h so Windows uses Sleep.  But
then I figured other platforms may have similarly restricted usleep
implementations, and the "correct" fix would be to use nanosleep for
delays instead.  Hence, this patch.

Note that MinGW doesn't currently have nanosleep, so Windows will still
fall back to using Sleep, but I've also tested this under Ubuntu 11 to
confirm the nanosleep implementation works.

The "do {...} while (0)" loop is used to "swallow the semicolon", see
http://gcc.gnu.org/onlinedocs/cpp/Swallowing-the-Semicolon.html