Seek to end of file before checking length when opening upload file
Unlike Linux, under Windows, the position returned by ftell() immediately
after opening a file for append is 0, regardless of the file size. A bug
was raised in MinGW because of this difference:

http://sourceforge.net/p/mingw/bugs/1558/

But it was closed as "won't fix" because Microsoft documents its ftell()
behaviour here:

http://msdn.microsoft.com/en-us/library/0ys3hc0b

Specifically, note this paragraph:

"Note that when a file is opened for appending data, the current file
position is determined by the last I/O operation, not by where the next
write would occur. For example, if a file is opened for an append and the
last operation was a read, the file position is the point where the next
read operation would start, not where the next write would start. (When a
file is opened for appending, the file position is moved to end of file
before any write operation.) If no I/O operation has yet occurred on a
file opened for appending, the file position is the beginning of the
file."

As such, to properly determine the pre-existing length of a file opened
for append under Windows, we must first perform a seek to the end of the
file.  Hence this change to add an fseek() call when checking the length
of the output file for an upload operation.