Report bcdDevice for found DFU devices
To provide additional information to users about their DFU device(s), the
bcdDevice field is now displayed as part of the DFU device information.
This field is used to specify the firmware version of the device, and the
DFU suffix contains a matching field which can be used for performing
version checking.

In addition, the print_dfu_if() function is now used to display device
information when first opening the DFU capable device, for consistency
with other locations where device information is displayed (and because
it is more important to know the bcdDevice version at this time, rather
than after having switched into DFU mode).