Fix crash on listing DFU interfaces
The recent change in 78d8efe to add serial number display introduced a
bug where the DFU list operation would crash if a DFU capable device was
found which had a serial number, but no alternate interface name string.
This is because the code to open the USB device, if it had not
previously been opened, was only done in the get_alt_name() function,
when an alternate interface name string is present, and not in the
get_serial() function.

Both get_alt_name() and get_serial() have been modified to open the USB
device, if necessary, and close it afterwards if it was not open when the
function was called.  In addition, the dev_handle structure element is now
being set to NULL after each close to avoid potential use of a bad handle.