DFU mode vendor, product, and serial specification
The "-d" option has been expanded to allow separate specification of the
vendor and product IDs for a device in DFU mode.  This enhancement was
done in a backwards compatible fashion such that anyone using the previous
"-d" functionality should see identical matching behaviour (i.e. same
vendor and product ID expected in both run-time and DFU mode).  To specify
DFU mode vendor and product IDs, add a comma to the end of the run-time
vendor and product ID specifier, followed by the DFU mode vendor and
product ID.  For example, "-d 1111:2222,3333:4444" will match only a
run-time device with vendor ID 0x1111 and product ID 0x2222, OR a DFU mode
device with vendor ID 0x3333 and product ID 0x4444.  Leaving any component
blank will match any device for that field (e.g. "-d 1111:,:4444" will
match a run-time device with vendor ID 0x1111 and ANY product ID, OR a DFU
mode device with ANY vendor ID but the product ID must be 0x4444).

The "-S" option has been expanded to allow separate specification of the
serial number for a device in DFU mode.  A DFU mode serial number is
specified by adding a comma following the run-time serial number, followed
by the DFU mode serial number.  If no comma is found in the argument then
the same serial number will be expected in both run-time and DFU mode.  If
a comma is present but no serial number is specified after the comma, then
any serial number is permitted for a device in DFU mode.  Note that comma
is not a valid character in a USB serial number, so this parsing shouldn't
conflict with the ability to specify any valid serial number.