Do not bail if GET_STATUS is not supported in appIDLE state
According to both versions of DFU specification(v. 1.1 and 1.0) it is
optional for a DFU compatible device to implement GET_STATUS request
in appIDLE state, which means that current behaviour of dfu-util,
namely to bail out if GET_STATUS request fails, is incompatible with
devices that do not implement aforementioned request in appIDLE state.
This commit adds the support of such devices by making assumption that
device is in appIDLE state if GET_STATUS returns -EPIPE and proceeding
to send DFU_DETACH request.

[Bradley Scott]
- remove old extraneous space in dfu_get_status() call
- initialize all fields of status structure
- only report state and status if it was retrieved successfully

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>
Reviewed-by: Bradley Scott <Bradley.Scott@confuseacat.org>
Tested-by: Bradley Scott <Bradley.Scott@confuseacat.org>
Signed-off-by: Bradley Scott <Bradley.Scott@confuseacat.org>