Rework device vendor/product matching
Changed parsing of device vendor and product ID filters to allow vendor
IDs to be skipped when specifying only product IDs, and to default the DFU
mode vendor and product IDs to "ANY" when not specified rather than
matching the specified runtime IDs.

Changed the way vendor ID, product ID, and serial number matching is done
during DFU device probing such that the DFU mode filters only apply to a
device in DFU mode while the runtime filters only apply to a device in
runtime mode.  This prevents undesired behaviour such as mixing the
runtime vendor ID with DFU mode product ID when matching a device.

After initiating DFU mode, the runtime mode match filters are set to
illegal values to look only for devices in DFU mode when attempting to
reconnect to the device.

If a runtime device vendor/product ID is not specified by the user during
a download operation, but the supplied DFU firmware file specifies vendor
and/or product IDs, then the IDs from the DFU firmware file will be used
as runtime mode match filters.

The checking of vendor and product IDs when downloading a DFU firmware
file has been adjusted to allow a match to either the runtime mode
vendor-product ID, or the DFU mode vendor-product ID, rather than only the
DFU mode vendor-product ID.  The DFU 1.1 specification recommends that the
DFU firmware files contain the runtime vendor and product ID values rather
than the DFU mode versions.

The bcdDevice version number has been re-added to the dfu-util --list
output.

A warning message that was mistakenly being treated as an error has now
been changed to use warnx() to be treated as a warning again.

Signed-off-by: Bradley Scott <Bradley.Scott@confuseacat.org>