Reconcile VID:PID and serial number specification
The parsing of the device vendor and product IDs and of serial numbers has
been reworked to make it easier to specify the device filter parameters
for a variety of expected use cases.  For vendor and product IDs, "*" is
now recognized as a "match any" operator, and "-" is recognized as a
"match none" operator.  For serial numbers, an empty serial number string
is now interpreted as "match any" to allow for a serial number to only be
specified for one of the modes (run-time or DFU) if desired.

A bug was corrected in the serial number filtering code that had swapped
the run-time and DFU mode versions of the serial number match string.

The dfu-util man page has been updated to reflect recent changes to
command line parameters.

Signed-off-by: Bradley Scott <Bradley.Scott@confuseacat.org>