dfuse: Warn, not quit when second-guessing the memory map
We do the best we can to decode the interface string even if
ST's specifications not always are followed (by themselves).
In a previous all-over errx/warnx change we accidentally went
from warning to erroring out in this case, so change it back.

Thanks to Adam Pearce for noticing the regression.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>