dfuse: Do not request DFU_ABORT if we can stay in dfuDNLOAD-IDLE
The DFU_ABORT request is used to transit back to dfuIDLE
after DFU_DNLOAD requests. If the next request is also
DFU_DNLOAD this is not necessary.

This commit removes the abort request from dfuse_special_command()
and performs the abort request explicitly in the cases where
the next request can be a DFU_UPLOAD command as well as after
finishing an upload or download session.

This allows us to avoid unneeded DFU_ABORT requests which
causes trouble with some ST STR750 ARM7 devices.

Thanks to Bernd Krumboeck for testing on STR750FV2.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>