dfuse: Also allow suffix when using --dfuse-address option
Leave dfu_has_suffix to the defalt MAYBE_SUFFIX value also in
this case.

Some --dfuse-address (-s) option modifiers can be useful when
using .dfu files, but the suffix still needs to be parsed for
the .dfu file to be recognized as valid for the device.

Of course, the address part of this option can not be used
in this case, but modifiers such as "leave" are useful.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>