dfuse: Wait until device is no more busy after special command
It has been seen that STM32L151VD devices can still be in DNBUSY
state on the second DFU_GETSTATUS request after download of a
special command. Therefore keep polling (and waiting) until it is
done. Otherwise a subsequent, premature DFU_DNLOAD can cause
the device to stall the pipe and USB errors appears.

Thanks to Adam Pearce for reporting the issue, providing many
patch suggestions and a ton of testing!

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>