dfu_prefix: Rewrite handling of prefix requirements
Like in the similar commit for dfu_suffix, there is no
need for the intermediate flags any longer, so this
simplifies the code a bit.

Also add logic to check there is a prefix of correct type
when needed and none when there should not be any.

Summary of logic:

Adding a prefix (-a) requires the specification of prefix type
(-L or -T). If -s is used to specify the Stellaris address, -T
is implied. There must be no existing prefix.

Deleting or checking for prefix will fail if there is no prefix,
or if the prefix does not match the optionally specified prefix
type.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>