dfu-suffix: Simplify flagging of need for suffix
The intermediate flags dfu_has_suffix and dfu_want_suffix
are not needed any longer and only makes the code harder
to read. The logic is simple: Require suffix to be present
if one is to be checked or deleted, and request suffix to
be added if, well, a suffix is added.

This commit should not change the code behaviour in any way.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>