main: Don't release interface before reset
We should not release the interface before resetting the bus. I.e.
Windows needs the interface handle to work with the device on which the
Reset IOCTL should be called to. At the end of the switch statement it
will be released anyway, which should be enough.

I tested dfu-util compiled with libusbx. See windows_usb.c:

https://github.com/libusbx/libusbx/blob/master/libusb/os/windows_usb.c#L3151

It peels out the winusb_handle from the previously cleared interface
handle (line 3119).

The current Windows API does not support any reset on USB since (IMHO) Win7.
Therefore everybody needs to use libusbk.dll (the successor of libusb-win32).
The downside of this is, that the implementation for sending the reset
IOCTL also depends on the interface handle.

https://github.com/codemercenary/libusbk/blob/trunk/libusbK/src/lusbk_bknd_libusbk.c#L818

Signed-off-by: Michael Grzeschik <mgr@pengutronix.de>