configure.ac: Drop AM_MAINTAINER_MODE
Thanks to Paul Fertser for pointing out it is not necessary:

Basically, non-maintainer mode means that autoconf doesn't create
rules to regenerate configure when configure.ac is changed. It's
supposed to benefit users who do not have autoconf installed but I can
hardly see how it can be so. If a user modifies configure.ac, he or
she clearly wants to have configure regenerated, so it's better to
have an error because of missing autoconf than to have the changes
silently ignored.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>