Assume DfuSe device is in DFU mode if bInterfaceProtocol is 0
The DFU interface descriptor has bInterfaceProtocol 2 if the device
is in DFU mode, and 1 in Runtime mode, according to the standard.
However, some devices have a zero instead. For now limited to DfuSe
devices (which usually only expose the DFU interface if they are in
DFU mode), we will also interpret a zero value as DFU mode.

Otherwise these devices would be thought to be in Runtime mode, in
which case dfu-util will not (after the recent Freerunner fix) apply
any alternate interface number/name filters, and if multiple alternate
interfaces are counted, dfu-util will refuse to carry on.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>