Check for DFU functional descriptor after checking for DFU class
I noticed that dfu-util was not recognising a DFU interface when it
was after a HID interface in the config descriptor. Turned out that the
Descriptor Type for a HID Descriptor and DFU Functional Descriptor are
the same (0x21). The find_descriptor() calls in probe_configuration()
function would therefore identify the HID interface as a possible DFU
interface before checking the Class and Subclass of this interface and
jump to found_dfu with the wrong interface. Moving the find_descriptor()
to after the Class/Subclass fixed this particular issue.