dfuse: Erase all needed pages before programming
Erase all relevant pages in a first pass, then do the programming
in a second pass (for each .dfu file element, or .bin file).

This is closer to what ST's DfuSe tool does, and it works around
some buggy bootloaders which don't know the real page size, for
example on the Radiolink AT9 RC transmitter.

https://sourceforge.net/p/dfu-util/tickets/20/

Signed-off-by: Geoffrey Hausheer <rc2012@pblue.org>
[Tormod] Only check for writeable page once
Signed-off-by: Tormod Volden <debian.tormod@gmail.com>