Implement our own get_string_descriptor_ascii() function
The one in libusb rejects string descriptors where the length field
(first byte) is wrong and has a greater value than the actual length
of the descriptor that is returned. We will instead fix up the length
field to the returned length.

This works around silicon bugs in STM32F427 (ROM bootloader).

https://sourceforge.net/p/dfu-util/tickets/22/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>