dfuse-pack.py: Fix indexing bug in multi-segment Intel HEX
The bug manifested itself on image segment 2 (and onward) for
multi-image targets, because tobinstr() was used once to extract
the whole data including padded holes between segments and these
holes were not taken into consideration in the index calculation.

The better solution is to use the start and end parameters of the
tobinstr() method, which nicely match (bar off-by-one) the starts
and ends returned from the segments() method.

https://sourceforge.net/p/dfu-util/tickets/35/

Signed-off-by: Colin Parker <colin.parker@aclima.io>